+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## 1.2.1 (2020-05-25)
 - Adopt CMake to compile SSP
 - Add script change_chrname_[to|from]_Greek.pl
+- Remove common/statistics.*
 
 ## 1.2.0 (2020-05-19)
 - Adopt htslib API for parsing SAM/BAM/CRAM format